feat: add default template component
