+++ src/components/template/template.styles.js
@@ -1,3 +1,4 @@
+import styled from 'styled-components';
 import { makeStyles } from '@material-ui/core/styles';
 
 export const useStyles = makeStyles((theme) => ({
@@ -17,3 +18,7 @@ export const useStyles = makeStyles((theme) => ({
     width: 'auto',
   },
 }));
+
+export const Wrapper = styled.div`
+  margin: ${({ theme }) => theme.spacing(2)}px;
+`;
